+++ packages/kofono/src/validator/expression/condition.spec.ts
@@ -9,6 +9,16 @@ import {
 } from "./condition";
 import type { Condition, Placeholder } from "./types";
 
+const t = `
+{property:my.selector:data} or
+{property:my.selector}
+{property:my.selector:validation}
+{property:my.selector:qualification}
+{property:my.selector}
+
+`;
+t;
+
 describe("parsePlaceholders()", () => {
     const tests: {
         template: string;
@@ -118,10 +128,10 @@ describe("testing var and def type", () => {
                 },
             },
             something: K.string().$q(q =>
-                q.condition("{var:user.role}", "==", "admin"),
+                q.expression("{var:user.role}", "==", "admin"),
             ),
             somethingElse: K.string().$q(q =>
-                q.condition("{var:user.role}", "==", "user"),
+                q.expression("{var:user.role}", "==", "user"),
             ),
         });
 
@@ -136,7 +146,7 @@ describe("testing var and def type", () => {
                         name: "bob",
                     },
                 })
-                .$q(q => q.condition("{def:test.name}", "==", "bob")),
+                .$q(q => q.expression("{def:test.name}", "==", "bob")),
         });
 
         expect(form.state.qualifications.something[0]).toBeTruthy();
@@ -148,10 +158,10 @@ describe("evaluateFieldValue()", () => {
         const form = await K.form({
             name: K.string().default("BOB"),
             something: K.string().$q(q =>
-                q.condition("{data:name|toLowerCase}", "==", "bob"),
+                q.expression("{data:name|toLowerCase}", "==", "bob"),
             ),
             somethingElse: K.string().$q(q =>
-                q.condition("{data:name}", "==", "bob"),
+                q.expression("{data:name}", "==", "bob"),
             ),
         });
 
@@ -175,11 +185,11 @@ describe("evaluateCondition()", () => {
                 .$v(v => v.required()),
             subscribeNewsletter: K.boolean()
                 .default(false)
-                .$q(q => q.condition("{data:acceptTerms}", "==", true)),
+                .$q(q => q.expression("{data:acceptTerms}", "==", true)),
             subscribeToMonthlyNewsletter: K.boolean()
                 .default(false)
                 .$q(q =>
-                    q.conditions([
+                    q.expressions([
                         ["{data:acceptTerms}", "==", true],
                         "and",
                         ["{data:subscribeNewsletter}", "==", true],