+++ packages/kofono/src/validator/schema.ts
@@ -1,16 +1,20 @@
 import type { SchemaAlphaValidator } from "./alpha/AlphaValidator";
 import type { SchemaAlphaNumValidator } from "./alphaNum/AlphaNumValidator";
 import type { SchemaBetweenValidator } from "./between/BetweenValidator";
-import type { SchemaConditionValidator } from "./condition/ConditionValidator";
 import type { SchemaDatetimeValidator } from "./datetime/DatetimeValidator";
 import type { SchemaEmailValidator } from "./email/EmailValidator";
 import type { SchemaEmptyValidator } from "./empty/EmptyValidator";
 import type { SchemaNotEmptyValidator } from "./empty/NotEmptyValidator";
 import type { SchemaEqualValidator } from "./equal/EqualValidator";
 import type { SchemaNotEqualValidator } from "./equal/NotEqualValidator";
+import type {
+    SchemaExpressionValidator
+} from "./expression/ExpressionValidator";
 import type { SchemaIfValidator } from "./if/IfValidator";
 import type { SchemaIncludesValidator } from "./includes/IncludesValidator";
-import type { SchemaNotIncludesValidator } from "./includes/NotIncludesValidator";
+import type {
+    SchemaNotIncludesValidator
+} from "./includes/NotIncludesValidator";
 import type { SchemaIsFalseValidator } from "./isFalse/IsFalseValidator";
 import type { SchemaIsTrueValidator } from "./isTrue/IsTrueValidator";
 import type { SchemaIsNotValidValidator } from "./isValid/IsNotValidValidator";
@@ -21,6 +25,7 @@ import type { SchemaMinValidator } from "./min/MinValidator";
 import type { SchemaPasswordValidator } from "./password/PasswordValidator";
 import type { SchemaRegexpValidator } from "./regexp/RegexpValidator";
 import type { SchemaRequiredValidator } from "./required/RequiredValidator";
+import type { SchemaNotSameAsValidator } from "./sameAs/NotSameAsValidator";
 import type { SchemaSameAsValidator } from "./sameAs/SameAsValidator";
 import type { SchemaUrlValidator } from "./url/UrlValidator";
 
@@ -32,11 +37,11 @@ export type SchemaPropertyValidator =
     | SchemaAlphaNumValidator
     | SchemaAlphaValidator
     | SchemaBetweenValidator
-    | SchemaConditionValidator
     | SchemaDatetimeValidator
     | SchemaEmailValidator
     | SchemaEmptyValidator
     | SchemaEqualValidator
+    | SchemaExpressionValidator
     | SchemaIfValidator
     | SchemaIncludesValidator
     | SchemaIsFalseValidator
@@ -49,6 +54,7 @@ export type SchemaPropertyValidator =
     | SchemaNotEmptyValidator
     | SchemaNotEqualValidator
     | SchemaNotIncludesValidator
+    | SchemaNotSameAsValidator
     | SchemaPasswordValidator
     | SchemaRegexpValidator
     | SchemaRequiredValidator