+++ packages/kofono/CHANGELOG.md
@@ -20,6 +20,7 @@
 - rename the wrong name ValidatorError enum to QualificationError
 - modify form stats compilation to round the progression value
 - add SameAsValidator
+- rename ConditionValidator to ExpressionValidator
 
 # 0.8.0 - 2026-04-28